Refactor out each language to it's own composite action
This makes it much easier to read the composite action, and reduces the need to include
the (increasingly complicated) `if:` check in each step. I'm not 100% sure this will work,
and I'm not aware of a way to test. I'm going to push this, then check that it works as
expected.