BREAKING CHANGE: Use explicit version for release (#24)
This was missed as part of https://github.com/pulumi/ci-mgmt/pull/900

We no longer use pulumictl to re-calculate the version multiple times
throughout a workflow. Instead, we use the
`pulumi/provider-version-action` to calculate the version, and prefer
passing the version around as an explicit variable to ensure it never
gets changed.

The new version is slightly different to the one generated by pulumictl
- using a shorthash of 7 characters rather than 8.

1. Add a new required input for `version` (BREAKING CHANGE).
2. Set the version input to the `PROVIDER_VERSION` which is the env name
we use in all other jobs for the version of the _provider_ - for the SDK
release sub-jobs which need it.
3. During the Java SDK publish sub-job, set the java-specific
`PACKAGE_VERSION` env variable.

## Breaking Change Rational

We always reference the exact version of this action and we've made
previous breaking changes, so it's better to avoid the complexity for
backward compatibility as we'll need to update all call sites with the
new version anyway.